updated the code that tempirarly disable the Done button to populate …

+++ Projects/Project1/app/src/main/java/com/example/mcresswell/project01/userProfile/UserProfile.java
@@ -50,7 +50,7 @@ public class UserProfile implements Parcelable {
                        String lifestyleSelection,
                        String weightGoal,
                        int lbsPerWeek,
-                       double weightInPounds,
+                       int weightInPounds,
                        int heightFeet,
                        int heightInches) {
